backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
cryptography==41.0.7
ecdsa==0.18.0
PyJWT==2.8.0
psycopg2-binary>=2.9.7
redis==5.0.1
python-dotenv==1.0.0
uuid==1.30
bcrypt==4.1.2
pytest==7.4.3
pytest-flask==1.3.0
gunicorn==21.2.0
SQLAlchemy==2.0.23 